Adicionado ocorrências ao mapa

+++ MataoMaps.Domain/Entities/Ocorrencia.cs
@@ -6,8 +6,8 @@ namespace MataoMaps.Domain.Entities
     public class Ocorrencia : EntityBase
     {
         public DateOnly Data { get; set; }
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
         public string FotoBase64 { get; set; }
         public string Endereco { get; set; }
         public string Descricao { get; set; }
@@ -22,8 +22,8 @@ namespace MataoMaps.Domain.Entities
 
         public Ocorrencia(
             DateOnly data,
-            decimal latitude,
-            decimal longitude,
+            double latitude,
+            double longitude,
             string fotoBase64,
             string endereco,
             string descricao,
